+++ src/main/java/fr/microtec/geo2/persistance/repository/common/GeoModifCorpsRepository.java
@@ -1,12 +1,10 @@
 package fr.microtec.geo2.persistance.repository.common;
 
-import java.math.BigDecimal;
-
 import org.springframework.stereotype.Repository;
 
 import fr.microtec.geo2.persistance.entity.common.GeoModifCorps;
 import fr.microtec.geo2.persistance.repository.GeoRepository;
 
 @Repository
-public interface GeoModifCorpsRepository extends GeoRepository<GeoModifCorps, BigDecimal> {
+public interface GeoModifCorpsRepository extends GeoRepository<GeoModifCorps, Integer> {
 }
